refactor: Move exceptions caused by internal errors to data.exception.internal
