+++ src/winter/controller/FrontController.java
@@ -22,12 +22,12 @@ import winter.data.client.JsonString;
 import winter.data.client.ModelView;
 import winter.data.enumdata.RequestVerb;
 import winter.data.exception.InvalidRequestVerbException;
-import winter.data.exception.InvalidReturnTypeException;
 import winter.data.exception.MappingNotFoundException;
 import winter.data.exception.annotation.AnnotationNotFoundException;
 import winter.data.exception.annotation.DuplicateMappingException;
 import winter.data.exception.initialization.InvalidPackageNameException;
 import winter.data.exception.initialization.PackageProviderNotFoundException;
+import winter.data.exception.internal.InvalidReturnTypeException;
 import winter.util.AnnotationScanner;
 import winter.util.ExceptionHandler;
 import winter.util.ReflectionUtil;
